[WIP] move to React Display field tab results
