+++ public/components/analyzer_form/custom.js
@@ -6,7 +6,7 @@ import {
   EuiTable,
   EuiTableBody
 } from '@elastic/eui';
-import displayRowsComponent from './abstract';
+import { displayRowsComponent } from './analyzer_form';
 
 class CustomAnalyzerComponent extends Component {
